requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
python-dotenv
